Remove NoopLauncher
It was added to support the now deprecated CopyArchiver.  However, it
introduces a requirement that the Node the parent project was built on
stay accessible through the execution of the join tasks.  This seems to
be causing more problems with other plugins, so dump the idea of the
NoopLauncher, and instead pass a 'null' launcher.